+++ R/sequenceFilter.R
@@ -7,10 +7,10 @@
 #' @param delta.time Numeric. Time between package downloads (seconds).
 #' @noRd
 
-sequenceFilter <- function(dat, packages, ymd, delta.time = 10) {
+sequenceFilter <- function(dat, p, ymd, delta.time = 240) {
   obs.versions <- unique(dat$version)
   if (length(obs.versions) > 1) {
-    history <- packageHistory(packages, check.package = FALSE)
+    history <- packageHistory(p, check.package = FALSE)
     sel <- history$Date <= ymd & history$Repository == "Archive"
     arch.pkg.history <- history[sel, ]
     removeSequences(dat, arch.pkg.history, delta.time = delta.time)
@@ -22,10 +22,10 @@ sequenceFilter <- function(dat, packages, ymd, delta.time = 10) {
 #' From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
 #' @param dat Object.
 #' @param arch.pkg.history Object.
-#' @param delta.time Numeric. Time between package downloads (seconds).
+#' @param delta.time Numeric. Time between first and last package download (seconds).
 #' @noRd
 
-removeSequences <- function(dat, arch.pkg.history, delta.time = 10) {
+removeSequences <- function(dat, arch.pkg.history, delta.time) {
   pkg.history <- arch.pkg.history
   history.obs <- nrow(pkg.history)
   
@@ -36,7 +36,6 @@ removeSequences <- function(dat, arch.pkg.history, delta.time = 10) {
 
     if (version.seq) {
       version.id <- seq_along(pkg.history$Version)
-      dat$date.time <- dateTime(dat$date, dat$time)
       dat <- dat[order(dat$date.time), ]
       
       runs <- rle(dat$version)
@@ -50,6 +49,8 @@ removeSequences <- function(dat, arch.pkg.history, delta.time = 10) {
       
       # single instance sequences
       candidates <- rle.data[rle.data$length == 1, ]
+      candidates <- candidates[!duplicated(candidates$ver), ]
+
       seq.start <- which(candidates$ver == pkg.history$Version[1])
       seq.stop <- which(candidates$ver == pkg.history$Version[history.obs])
       
@@ -69,14 +70,12 @@ removeSequences <- function(dat, arch.pkg.history, delta.time = 10) {
             start.stop <- rle.data[seq.tmp.obs, ]
             obs.chk <- unique(unlist(start.stop[, c("start", "stop")]))
             tmp <- dat[obs.chk, ]
-            tmp$date.time <- dateTime(tmp$date, tmp$time)
             time.range <- range(tmp$date.time)
             time.window <- delta.time * nrow(tmp)
             time.range.delta <- difftime(time.range[2], time.range[1],
               units = "sec")
             if (time.range.delta < time.window) obs.chk
           }))
-        
           obs.exclude <- row.names(dat[rle.exclude, ])
         }
       } 
@@ -91,7 +90,6 @@ removeSequences <- function(dat, arch.pkg.history, delta.time = 10) {
         candidate <- candidate[candidate$version %in% pkg.history$Version, ]
         all.archive.vers <- all(pkg.history$Version %in% candidate$version)
         
-        candidate$date.time <- dateTime(candidate$date, candidate$time)
         candidate <- candidate[order(candidate$date.time), ]
         
         seq.start <- candidate[candidate$version == first.pkg.version, ]